+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 ### Fixed
 - Publish Android artifacts for `debug` variant
+- Fix IndexOutOfBoundsException in texture builder
 
 ## [1.0.0-ALPHA1]
 ### Added